+++ src/database/user/auth/error/mod.rs
@@ -1,2 +1,4 @@
-mod user_auth;
 pub use user_auth::UserAuthError;
+
+mod user_auth;
+
