+++ src/database/user/auth/error/user_auth.rs
