--- src/database/user/token/mod.rs
@@ -1,7 +1,0 @@
-mod claims;
-mod decode;
-pub mod local;
-pub mod error;
-
-pub use claims::TokenClaims;
-pub use decode::decode;
