+++ src/protocol/mod.rs
@@ -1,5 +1,6 @@
+pub use packet::{PacketClient, PacketResult, PacketServer};
+
 mod handler;
 mod packet;
 mod structure;
 
-pub use packet::{PacketClient, PacketResult, PacketServer};
