+++ src/protocol/structure/mod.rs
@@ -1,4 +1,5 @@
+pub use common::*;
+
 mod common;
 mod wordchain;
 
-pub use common::*;
