+++ src/websocket.rs
@@ -1,10 +1,11 @@
+use std::time::{Duration, Instant};
+
 use actix::*;
-use actix_web::{web, Error, HttpRequest, HttpResponse};
+use actix_web::{Error, HttpRequest, HttpResponse, web};
 use actix_web_actors::ws;
+use log::{info, warn};
 
 use crate::protocol::{PacketClient, PacketServer};
-use log::{info, warn};
-use std::time::{Duration, Instant};
 
 const HEARTBEAT_INTERVAL: Duration = Duration::from_secs(5);
 const CLIENT_TIMEOUT: Duration = Duration::from_secs(10);
